requirements.txt
amqp==5.2.0
asgiref==3.6.0
attrs==23.2.0
billiard==4.2.0
Brotli==1.1.0
celery==5.3.6
certifi==2023.11.17
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
contourpy==1.2.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==41.0.7
cssselect2==0.7.0
cvzone==1.6.1
cycler==0.12.1
defusedxml==0.8.0rc2
Django==3.2.23
django-cors-headers==4.0.0
django-currentuser==0.5.3
django-filter==23.5
django-rest-swagger==2.2.0
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
djoser==2.2.2
docopt==0.6.2
drf-extra-fields==3.6.1
drf-spectacular==0.27.0
drf-yasg==1.21.7
elastic-transport==8.11.0
elasticsearch==8.11.1
elasticsearch-dsl==8.11.0
environs==9.5.0
filetype==1.2.0
fonttools==4.47.2
fuzzywuzzy==0.18.0
gunicorn==20.1.0
html5lib==1.1
idna==3.6
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.3
jsonschema==4.21.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
kombu==5.3.5
Levenshtein==0.23.0
lxml==4.9.3
Markdown==3.5.2
markdown2==2.4.12
MarkupSafe==2.1.3
marshmallow==3.20.2
matplotlib==3.8.2
md2pdf==1.0.1
mistune==3.0.2
mysqlclient==2.2.1
numpy==1.26.3
oauthlib==3.2.2
openapi-codec==1.3.2
opencv-python==4.9.0.80
packaging==23.2
pandas==2.1.4
pillow==10.2.0
prompt-toolkit==3.0.43
pycparser==2.21
pydyf==0.8.0
PyJWT==2.3.0
pyparsing==3.1.1
pyphen==0.14.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-Levenshtein==0.23.0
python3-openid==3.2.0
pytz==2023.3.post1
PyYAML==6.0.1
rapidfuzz==3.6.1
redis==5.0.1
referencing==0.32.1
reportlab==4.0.9
requests==2.31.0
requests-oauthlib==1.3.1
rpds-py==0.17.1
seaborn==0.13.1
simplejson==3.19.1
six==1.16.0
social-auth-app-django==5.4.0
social-auth-core==4.4.2
sqlparse==0.4.4
tinycss2==1.2.1
tqdm==4.66.1
typing_extensions==4.9.0
tzdata==2023.4
uritemplate==4.1.1
urllib3==2.1.0
vine==5.1.0
wcwidth==0.2.13
weasyprint==60.2
webencodings==0.5.1
whitenoise==6.4.0
yandexgptlite==0.1.7
zopfli==0.2.3